predict_scan.py edited online with Bitbucket

+++ models/patches_v2/predict_scan.py
@@ -119,8 +119,8 @@ def predict_case(case):
     preds = np.vstack(preds)
     
     # Reshape the image to the original shape, so we can map predictions to actual locations
-    siz1 = ((img.shape[0]-int(patch_size/2))/scan_step+1)
-    siz2 = ((img.shape[1]-int(patch_size/2))/scan_step+1)
+    siz1 = (int((img.shape[0]-int(patch_size/2)))/scan_step+1)
+    siz2 = (int((img.shape[1]-int(patch_size/2)))/scan_step+1)
     preds = preds.reshape([siz1,siz2,preds.shape[1]])
     return preds, time.time()-t1
